+++ MaquetaParaFinal/Clases/Agregar/VentanaAgregarPractica.cs
@@ -39,21 +39,6 @@ namespace MaquetaParaFinal.View.Agregar
                 txtNombrePractica.Text = "Nombre";
             }
         }
-        private void ControlarNombre(object sender, TextChangedEventArgs e)
-        {
-            TextBox textBox = (TextBox)sender;
-            string input = textBox.Text;
-
-            // Patrón para permitir letras, espacios y comilla simple
-            string regEx = @"^[A-Za-z ']{1,20}$";
-
-            if (!(Regex.IsMatch(input, regEx) && input.Length <= 20)) // La entrada no cumple con el patrón, elimina caracteres no válidos
-            {
-                textBox.Text = Regex.Replace(input, @"[^A-Za-z ']", "");
-                textBox.Text = textBox.Text.Substring(0, Math.Min(20, textBox.Text.Length)); // Limita a 20 caracteres
-                textBox.Select(textBox.Text.Length, 0); // Coloca el cursor al final del texto
-            }
-        }
 
         private void btnCancelar_Click(object sender, RoutedEventArgs e) => this.Close();
 
@@ -123,7 +108,22 @@ namespace MaquetaParaFinal.View.Agregar
 
         }
 
-        private void SoloNumero_PreviewTextInput(object sender, TextCompositionEventArgs e)
+        private void ControlarNombre(object sender, TextChangedEventArgs e)
+        {
+            TextBox textBox = (TextBox)sender;
+            string input = textBox.Text;
+
+            string regEx = @"^[A-Za-z ']{1,20}$";
+
+            if (!(Regex.IsMatch(input, regEx) && input.Length <= 20)) // La entrada no cumple con el patrón, elimina caracteres no válidos
+            {
+                textBox.Text = Regex.Replace(input, @"[^A-Za-z ']", "");
+                textBox.Text = textBox.Text.Substring(0, Math.Min(20, textBox.Text.Length)); // Limita a 20 caracteres
+                textBox.Select(textBox.Text.Length, 0); // Coloca el cursor al final del texto
+            }
+        }
+
+        private void SoloNumeros_PreviewTextInput(object sender, TextCompositionEventArgs e)
         {
             if (IsNumber(e.Text)) e.Handled = true;
         }
@@ -136,5 +136,6 @@ namespace MaquetaParaFinal.View.Agregar
             }
             return false;
         }
+
     }
 }
