+++ MaquetaParaFinal/View/Agregar/AgregarMedico.xaml.cs
@@ -31,18 +31,5 @@ namespace MaquetaParaFinal.View.Agregar
             CargarServicios();
         }
 
-        private void SoloNumeros_PreviewTextInput(object sender, TextCompositionEventArgs e)
-        {
-            if (IsNumber(e.Text)) e.Handled = true;
-        }
-
-        private bool IsNumber(string text)
-        {
-            foreach (char c in text)
-            {
-                if (!char.IsDigit(c)) return true;
-            }
-            return false;
-        }
     }
 }
