+++ MaquetaParaFinal/View/Agregar/AgregarServicio.xaml
@@ -18,7 +18,7 @@
                 <DockPanel Margin="10,20,10,0">
                         <Label Content="Nombre*" Margin="0,3,0,0" Width="100" FontWeight="Medium" FontFamily="Roboto" Foreground="#F0F0F0"/>
                         <Border Margin="0,6,10,0"  HorizontalAlignment="Stretch" Background="#213555" CornerRadius="9">
-                            <TextBox x:Name="txtNombreServicio" Margin="5,0,0,0" TextWrapping="Wrap" Text="Nombre" VerticalAlignment="Top" Background="Transparent" Foreground="#F0F0F0" BorderBrush="Transparent" LostFocus="txtNombreServicio_LostFocus" GotFocus="txtNombreServicio_GotFocus" />
+                            <TextBox x:Name="txtNombreServicio" Margin="5,0,0,0" TextWrapping="Wrap" Text="Nombre" VerticalAlignment="Top" Background="Transparent" Foreground="#F0F0F0" BorderBrush="Transparent" LostFocus="txtNombreServicio_LostFocus" GotFocus="txtNombreServicio_GotFocus" TextChanged="ControlarNombre" />
                         </Border>
                  </DockPanel>
                  <DockPanel Width="250" Height="35" Margin="0,25,0,5">
